@@ -1,8 +1,6 @@
 module github.com/loft-sh/cluster-api-provider-vcluster
 
-go 1.21
-
-toolchain go1.22.0
+go 1.22
 
 require (
 	github.com/golangci/golangci-lint v1.56.2
